requirements.txt
openai>=1.68.0
anthropic>=0.49.0
python-dotenv>=1.0.1
swiftapi-python>=1.0.3
